#include <11.h>.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  * of the number 612852475143 using the math library to optimize.
  * Return: 0 (Success)
  */
-int main(void)
+long main(void)
 {
 	long num = 612852475143;
 	long largestPrimeFactor = -1;
